refac: separate structs for each channel
doc: update examples in readme
style: format code

Closes #3
